# spec/openapi.yaml
openapi: 3.1.0
info:
  title: HR Platform API
  version: 1.0.0-alpha.1
  description: |
    Multi-tenant HR platform API.

    **This specification is the contract.** Server interfaces and the Kotlin, Swift and TypeScript
    clients are all generated from it, and a breaking change to a published version fails CI.
    See `docs/03-architecture.md` §9.

    ## Tenancy

    Every request is resolved to exactly one tenant. Precedence, highest first:

    1. The `tenant_id` claim in the bearer token — authoritative once authenticated.
    2. The `X-Tenant-Code` header — for the unauthenticated login and org-resolve endpoints.
    3. The request subdomain.

    Supplying a token and a conflicting `X-Tenant-Code` is rejected with `TENANT_MISMATCH`.

    ## Errors

    Every non-2xx response uses one envelope. Clients localise from the machine-readable `code`,
    never from `message` — the server cannot reliably know the caller's locale, and we ship in
    six languages.

    ## Pagination

    Cursor-based everywhere. Offset pagination is not offered on tenant-scoped collections: it
    re-scans skipped rows and silently drops and duplicates records on concurrently-written
    tables such as attendance and audit.

    ## Idempotency

    Every unsafe method accepts an `Idempotency-Key` header and it is **required** for endpoints
    the mobile outbox can replay. A repeat of a key already processed returns the original
    response rather than applying the operation twice.
...
servers:
...
security:
  - bearerAuth: []
paths:
  /v1/auth/resolve-tenant:
    post:
      tags: [Authentication]
      summary: Resolve an organisation from a work email or org code
      description: |
        The first step of sign-in. Deliberately does **not** reveal whether a user account exists —
        it resolves the organisation only, so this endpoint cannot be used to enumerate accounts.

        This replaces asking the user to type a "service URL", which is a common source of
        support load in comparable products.
      operationId: resolveTenant
      security: []
      parameters:
        # Required: the tenant is resolved by the server's request filter from this header (or
        # from the subdomain). Send the same value as `orgCode` in the body.
        - $ref: '#/components/parameters/TenantCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveTenantRequest'
      responses:
        '200':
          description: Organisation resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResolveTenantResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/auth/token:
    post:
      tags: [Authentication]
      summary: Sign in with a password
      description: |
        Exchanges credentials for an access and refresh token pair, and registers the calling
        device in the same round trip so the client can offer biometric enrolment immediately.

        Failures are deliberately uniform: a wrong password and an unknown username both return
        `INVALID_CREDENTIALS`, so this endpoint cannot be used to enumerate accounts. Repeated
        failures lock the account per the tenant's password policy.
      operationId: issueToken
      security: []
      parameters:
        - $ref: '#/components/parameters/TenantCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordGrantRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: |
            Invalid credentials, or MFA is required. `MFA_REQUIRED` carries an `mfaToken` in
            `details` which must be presented to `/v1/auth/mfa/verify`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/auth/token/refresh:
    post:
      tags: [Authentication]
      summary: Exchange a refresh token for a new access token
      description: |
        Refresh tokens are single-use and rotate on every exchange.

        If a token that has already been used is presented again, the most likely explanation is
        theft, so the entire token family is revoked and every device sharing that login session
        must re-authenticate. The response is `TOKEN_REUSE_DETECTED`.

        Requires `X-Tenant-Code`: the token lookup is itself tenant-scoped by row-level security,
        so the tenant must be known before the token can be found.
      operationId: refreshToken
      security: []
      parameters:
        - $ref: '#/components/parameters/TenantCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Rotated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/auth/token/biometric:
    post:
      tags: [Authentication]
      summary: Exchange a device-sealed refresh token after a biometric assertion
      description: |
        The core of the "no password after enrolment" experience.

        The refresh token is sealed at enrolment in the device's Android Keystore or iOS Secure
        Enclave under a key that requires user authentication. The OS releases it only after a
        successful fingerprint or face match, and invalidates the key entirely if the device's
        enrolled biometrics change.

        The server therefore trusts the *presentation* of the sealed token as evidence that the
        biometric check passed — it never receives or stores biometric data itself.

        Two conditions beyond a normal refresh: the token must have been issued to the device
        presenting it, and that device must have completed biometric enrolment. Together these
        stop a token exfiltrated from one device being replayed from another.
      operationId: biometricToken
      security: []
      parameters:
        - $ref: '#/components/parameters/TenantCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BiometricGrantRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Device revoked, or re-authentication with a password is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /v1/auth/devices:
    get:
      tags: [Authentication]
      summary: List the current user's registered devices
      description: |
        Powers the "where am I signed in?" screen. The device making the request is flagged with
        `current: true` so the client can avoid offering the user a way to revoke themselves
        without warning.
      operationId: listDevices
      responses:
        '200':
          description: Devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
    post:
      tags: [Authentication]
      summary: Register the current device
      description: |
        Registers or updates a device for the authenticated user, including its push token.

        Called on sign-in and again whenever the push token rotates. Idempotent on
        `(user, deviceId)`: re-registering an existing device updates it rather than creating a
        duplicate.
      operationId: registerDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceRequest'
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
  /v1/auth/devices/{id}:
    delete:
      tags: [Authentication]
      summary: Revoke a device
      description: |
        Revokes the device and every refresh token bound to it. Takes effect immediately for
        refresh and biometric grants; an access token already issued remains valid until it
        expires, at most fifteen minutes.

        Note the parameter is the device record's `id` (a UUID), not its `deviceId` (the
        client-generated string). The two are distinct and naming the path variable `deviceId`
        was actively misleading.
      operationId: revokeDevice
      parameters:
        - name: id
          in: path
          required: true
          description: The device record's UUID, as returned by `GET /v1/auth/devices`.
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Revoked
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/auth/logout:
    post:
      tags: [Authentication]
      summary: Sign out
      description: |
        Revokes the token family the presented refresh token belongs to. Other devices stay signed
        in — signing a user out everywhere is a separate, deliberate action.

        The body is optional. Without it the caller's entire set of sessions is revoked, which is
        the correct fallback when a client has lost track of its refresh token.

        Never fails: sign-out is idempotent and a client must always be able to end its session.
      operationId: logout
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '204':
          description: Signed out
...
  /v1/reference:
    get:
      tags: [Reference]
      summary: Several reference taxonomies in one call
      description: |
        The mobile clients sync these as a unit and cache them indefinitely — they change perhaps
        twice a year. Serving them individually would mean four round trips to render one form,
        which is the chattiness that makes an app feel slow regardless of how fast each call is.
      operationId: listReferenceData
      parameters:
        - name: tables
          in: query
          description: Kebab-case taxonomy names, e.g. `blood-group`. Omit to fetch all of them.
          schema:
            type: array
            items: { type: string }
          style: form
          explode: false
      responses:
        '200':
          description: Taxonomy name to its items
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: '#/components/schemas/ReferenceItem'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/reference/{table}:
    get:
      tags: [Reference]
      summary: One reference taxonomy
      description: Kebab-case taxonomy name, e.g. `employee-category`.
      operationId: listReferenceTable
      parameters:
        - name: table
          in: path
          required: true
          schema: { type: string }
        - name: includeInactive
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReferenceItem'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `/v1/auth/token`. Lifetime 15 minutes.
  parameters:
    TenantCode:
      name: X-Tenant-Code
      in: header
      required: true
      description: Organisation code. Required on unauthenticated endpoints, where no token exists yet.
      schema:
        type: string
        pattern: '^[a-z0-9][a-z0-9-]{1,62}[a-z0-9]$'
...
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
    ApiErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable identifier. Clients localise from this.
          examples: [LEAVE_BALANCE_INSUFFICIENT]
        message:
          type: string
          description: Developer-facing English description. Not for display to end users.
        field:
          type: string
          description: Input field the error is attributable to.
        details:
          type: object
          additionalProperties: true
          description: Structured context for building a localised message.
          examples: [{ available: 2.5, requested: 5 }]
        violations:
          type: array
          items:
            $ref: '#/components/schemas/FieldViolation'
        requestId:
          type: string
          description: Correlates with server logs. Quote this to support.
    FieldViolation:
      type: object
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
        rejectedValue: {}
    ResolveTenantRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        orgCode:
          type: string
      description: Supply exactly one of `email` or `orgCode`.
    ResolveTenantResponse:
      type: object
      required: [code, name, authMethods]
      properties:
        code: { type: string }
        name: { type: string }
        logoUrl: { type: string, format: uri }
        brandColor: { type: string, pattern: '^#[0-9A-Fa-f]{6}$' }
        locale: { type: string }
        authMethods:
          type: array
          description: Which sign-in methods this organisation offers, so the client renders the right form.
          items:
            type: string
            enum: [PASSWORD, SSO_SAML, SSO_OIDC]
        ssoRedirectUrl: { type: string, format: uri }
    PasswordGrantRequest:
      type: object
      required: [username, password, device]
      properties:
        username: { type: string }
        password: { type: string, format: password }
        device:
          $ref: '#/components/schemas/DeviceInfo'
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
    BiometricGrantRequest:
      type: object
      required: [sealedRefreshToken, deviceId]
      properties:
        sealedRefreshToken:
          type: string
          description: Refresh token released from device secure hardware after a successful biometric assertion.
        deviceId:
          type: string
          description: Must match the device the token was bound to at enrolment.
    TokenResponse:
      type: object
      required: [accessToken, refreshToken, tokenType, expiresIn]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        tokenType: { type: string, enum: [Bearer] }
        expiresIn: { type: integer, description: Access token lifetime in seconds. }
        refreshExpiresIn: { type: integer }
        biometricEnrolmentOffered:
          type: boolean
          description: |
            True when this device has not yet enrolled biometrics. The client should offer
            enrolment immediately after first sign-in — deferring it is why users end up typing
            passwords forever.
    DeviceInfo:
      type: object
      required: [deviceId, platform]
      properties:
        deviceId: { type: string, maxLength: 128 }
        platform: { type: string, enum: [ANDROID, IOS, WEB, KIOSK] }
        model: { type: string }
        osVersion: { type: string }
        appVersion: { type: string }
        pushToken: { type: string }
    RegisterDeviceRequest:
      allOf:
        - $ref: '#/components/schemas/DeviceInfo'
        - type: object
          properties:
            attestation:
              type: string
              description: Play Integrity or App Attest token, verified server-side.
    Device:
      type: object
      required: [id, deviceId, platform, trusted]
      properties:
        id: { type: string, format: uuid }
        deviceId: { type: string }
        platform: { type: string, enum: [ANDROID, IOS, WEB, KIOSK] }
        model: { type: string }
        osVersion: { type: string }
        appVersion: { type: string }
        biometricEnrolled: { type: boolean }
        attestationVerified: { type: boolean }
        trusted: { type: boolean }
        lastSeenAt: { type: string, format: date-time }
        current: { type: boolean, description: True for the device making this request. }
...
    ReferenceItem:
      type: object
      required: [id, code, name, sequence, active]
      description: A row in a configurable lookup list — blood group, marital status, and so on.
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        description: { type: string }
        sequence:
          type: integer
          description: Display order. Lower first.
        active:
          type: boolean
          description: Inactive items are retained so historical records still resolve, but are not offered for new entry.
...